feat: DAO 쿼리 작성
